requirements.txt
googletrans==2.4.0
pillow==6.2.1
pytesseract=0.3.0
pyscreenshot=0.5.1
